+++ egucepayment-restapi/src/main/java/org/guce/epayment/core/repositories/SittingDateRepository.java
@@ -1,6 +1,5 @@
 package org.guce.epayment.core.repositories;
 
-import java.math.BigDecimal;
 import java.util.List;
 import org.guce.epayment.core.entities.SittingDate;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
  *
  * @author tadzotsa
  */
-public interface SittingDateRepository extends JpaRepository<SittingDate, BigDecimal> {
+public interface SittingDateRepository extends JpaRepository<SittingDate, Integer> {
 
     @Query("SELECT sd FROM SittingDate sd ORDER BY sd.insertionDate DESC")
     @Override
